pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vector4HN"
version = "0.1.0"
description = "Hacker News Deep Dive Assistant"
authors = [
    {name = "Giovanni Battista Pernazza"}
]
dependencies = [
    "httpx",
    "beautifulsoup4",
    "textual",
    "ollama",
    "python-dotenv",
    "markdownify",
    "trafilatura",
    "google-genai",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
vector = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
